Update the test directory to match template and move in real Cargo.toml contents
