+++ ide/Cargo.toml
@@ -1,12 +1,14 @@
-# You must change these to your own details.
+
 [package]
-name = "rust-webpack-template"
-description = "My super awesome Rust, WebAssembly, and Webpack project!"
+name = "flowide"
 version = "0.1.0"
-authors = ["You <you@example.com>"]
-categories = ["wasm"]
-readme = "README.md"
+authors = ["Andrew Mackenzie"]
 edition = "2018"
+categories = ["wasm"]
+description = "A web version of the 'flow' project for data driven parallel applications"
+license = "Apache-2.0/MIT"
+readme = "README.md"
+repository = "https://github.com/andrewdmackenzie/flow"
 
 [lib]
 crate-type = ["cdylib"]
@@ -17,24 +19,47 @@ crate-type = ["cdylib"]
 lto = true
 
 [features]
-# If you uncomment this line, it will enable `wee_alloc`:
-#default = ["wee_alloc"]
+default = ["console_error_panic_hook", "wasm-bindgen"]
 
 [dependencies]
 # The `wasm-bindgen` crate provides the bare minimum functionality needed
 # to interact with JavaScript.
-wasm-bindgen = "0.2.45"
+wasm-bindgen = {version = "0.2.45", optional = true }
+flowclib = { path = "../flowclib", version = "~0.7.0" }
+flowrlib = { path = "../flowrlib", version = "~0.7.0" }
+flowstdlib = { path = "../flowstdlib", version = "~0.7.0" }
+nodeprovider = { path = "../nodeprovider", version = "0.1.0" }
+serde_json = "1.0"
+wasm-logger = "0.1.4"
+log = {version = "0.4.6", features = ["std"]}
+
+# The `console_error_panic_hook` crate provides better debugging of panics by
+# logging them with `console.error`. This is great for development, but requires
+# all the `std::fmt` and `std::panicking` infrastructure, so isn't great for
+# code size when deploying.
+console_error_panic_hook = { version = "0.1.5", optional = true }
 
 # `wee_alloc` is a tiny allocator for wasm that is only ~1K in code size
-# compared to the default allocator's ~10K. However, it is slower than the default
-# allocator, so it's not enabled by default.
+# compared to the default allocator's ~10K. It is slower than the default
+# allocator, however.
 wee_alloc = { version = "0.4.2", optional = true }
 
 # The `web-sys` crate allows you to interact with the various browser APIs,
 # like the DOM.
 [dependencies.web-sys]
 version = "0.3.22"
-features = ["console"]
+features = [
+  "Document",
+  "Element",
+  'HtmlButtonElement',
+  'HtmlSelectElement',
+  'HtmlOptionsCollection',
+  "HtmlElement",
+  "Node",
+  "Text",
+  "Window",
+  "console",
+]
 
 # The `console_error_panic_hook` crate provides better debugging of panics by
 # logging them with `console.error`. This is great for development, but requires
